Add automatic PKGBUILD to release process
This adds an automatic self referential PKGBUILD to the release process.
When a release is cut, it will template in necessary details to a new
PKGBUILD. Arch users then only need to download this PKGBUILD and it
will handle downloading the linux release (which doesn't contain
necessary libraries), install libraries, and slot the release assets in
where necessary.